refactor: split ui/menus/feats into focused submodules
Separate requirements display, subchoice pickers, creation flow, and
level-up flow; keep public API on ui.menus.feats package.
