+++ ui/menus/feats/_constants.py
@@ -1,0 +1,18 @@
+"""Константы UI выбора черт."""
+
+ELEMENTAL_DAMAGE_TYPES = [
+    "acid",
+    "cold",
+    "fire",
+    "lightning",
+    "thunder",
+]
+
+MAGIC_INITIATE_CLASSES = [
+    "bard",
+    "cleric",
+    "druid",
+    "sorcerer",
+    "warlock",
+    "wizard",
+]
